Changed some things in PartFactory
Changed partsListFile into a String, opening a file should be done in a
try block, makePartsFromFile() now uses a constant to find its file. In
the catch block removed the new thrown RunTimeException, we now print
out the error and return an empty array list so the program continues to
run.